requirements.txt
langchain_openai
langchain
fastapi
openai
uvicorn
vaderSentiment
python-dotenv